Second queue on matches SNS topic to increment pg counters
